fix: 피드 모달 위치 BaseLayout 으로 이동
