+++ src/components/profile/ProfileFeedList.tsx
@@ -1,14 +1,10 @@
 import React from 'react';
 import { useRouter } from 'next/router';
-import { useRecoilState, useSetRecoilState } from 'recoil';
 import { FeedType } from '@/core/types/feed';
 import FeedService from '@/services/feed';
 import useAuth from '@/hooks/useAuth';
-import { feedModalState, isFeedModalOpenState } from '@/store/feedAtom';
 import InfinityDataList from '../common/InfinityDataList';
-import FeedModal from '../common/FeedModal';
 import ProfileFeedListItem from './ProfileFeedListItem';
-import ProfileFeedModal from './ProfileFeedModal';
 
 interface ProfileFeedListProps {
   queryKey: string;
@@ -18,15 +14,6 @@ function ProfileFeedList({ queryKey }: ProfileFeedListProps) {
   const router = useRouter();
   const { username } = router.query;
   const { payload } = useAuth();
-
-  const setFeedModalState = useSetRecoilState(feedModalState);
-  const [isFeedModalOpen, setIsFeedModalOpen] =
-    useRecoilState(isFeedModalOpenState);
-
-  const onClickFeedModalOpenHandler = () => {
-    setIsFeedModalOpen(false);
-    setFeedModalState(null);
-  };
 
   return (
     <>
@@ -52,12 +39,6 @@ function ProfileFeedList({ queryKey }: ProfileFeedListProps) {
           propsObject={{ queryKey: [`${queryKey}-${username}`] }}
         ></InfinityDataList>
       </div>
-      <FeedModal
-        isModalOpen={isFeedModalOpen}
-        onClickCloseModal={onClickFeedModalOpenHandler}
-      >
-        <ProfileFeedModal />
-      </FeedModal>
     </>
   );
 }
